Blink IDL data update using latest JSONification code
